Downgrade josekit version (0.10.3->0.7.4)
As a general rule, in case an upstream project needs
to be packaged (i.e.: Fedora), it should be checked
that packaging versions are available. Latest `josekit`
version in Fedora Rawhide is 0.7.4. I have checked
this does not affect compilation/test execution, so,
ideally, package needs to be downgraded so that future
package can be added to Fedora

Signed-off-by: Sergio Arroutbi <sarroutb@redhat.com>